+++ Pipelines/src/Pipelines.jl
@@ -16,6 +16,8 @@ export Card,
 
 export get_metadata, register_card, CardConfig
 
+public SourceVariables, OutputVariables, get_node_inputs, get_node_outputs
+
 public train!, evaljoin, train_evaljoin!
 
 public train, evaluate, inputs, outputs, invertible